requirements.txt
django-oauth-toolkit==1.5.0
python-social-auth==0.3.6
drf-social-oauth2==1.0.9
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
python-social-auth==0.3.6
social-auth-app-django==4.0.0
drf-social-oauth2==1.0.9
django-cors-headers==3.7.0